Adding security check for deleting eating plan

+++ app/Http/Controllers/EatingPlanController.php
@@ -103,7 +103,9 @@ class EatingPlanController extends Controller
      */
     public function destroy(EatingPlan $eatingplan)
     {
-
+        if ($eatingplan->nutritionist->user != auth()->user()) {
+            return response('', 403);
+        }
         $patient = $eatingplan->patient;
         $eatingplan->delete();
